refactor: renamed PyConv resnet module
