+++ holocron/models/classification/__init__.py
@@ -4,7 +4,7 @@ from .darknetv2 import *
 from .darknetv3 import *
 from .darknetv4 import *
 from .mobileone import *
-from .pyconvresnet import *
+from .pyconv_resnet import *
 from .repvgg import *
 from .res2net import *
 from .resnet import *
